requirements.txt
beautifulsoup4==4.12.3
Ghost.py==0.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.30
requests==2.31.0
selenium==4.20.0
solana==0.34.0
solders==0.21.0
Telethon==1.35.0
django==5.0.4
djangorestframework==3.15.1
pyppeteer==2.0.0
alembic==1.13.1
aioconsole==0.7.1
black==24.4.2
django-configurations==2.5.1